Change contract of StubFactory
Instead of an instance method we now use a static supports() method
to determine if a factory matches a given stub type. This is better
for the lifecycle since bean definitions have to be created before
and factory can actually be created and autowired.

It's a slightly unusual contract, but it makes it much harder to
make a mistake with lifecycle, and much easier to install custom
StubFactory instances (as @Beans)